+++ packages/components/atoms/f-button/test/visual/f-button.visual.desktop.spec.js
@@ -1,4 +1,3 @@
-const { buildUrl } = require('@justeat/f-wdio-utils/src/storybook-extensions');
 const Button = require('../../test-utils/component-objects/f-button.component');
 
 let button;
@@ -10,11 +9,9 @@ describe('f-button Desktop visual tests', () => {
             button = new Button();
             button.withQuery('knob-Button Type', 'primary');
             button.withQuery('knob-Button Size', 'medium');
-            const pageUrl = buildUrl(button.componentType, button.componentName, button.path);
 
             // Act
-            button.open(pageUrl);
-            button.waitForActionComponent();
+            button.load();
 
             // Assert
             browser.percyScreenshot('f-button - Primary - Medium', 'desktop');
@@ -25,11 +22,9 @@ describe('f-button Desktop visual tests', () => {
             button = new Button();
             button.withQuery('knob-Button Type', 'primary');
             button.withQuery('knob-Button Size', 'large');
-            const pageUrl = buildUrl(button.componentType, button.componentName, button.path);
 
             // Act
-            button.open(pageUrl);
-            button.waitForActionComponent();
+            button.load();
 
             // Assert
             browser.percyScreenshot('f-button - Primary - Large', 'desktop');
@@ -40,11 +35,9 @@ describe('f-button Desktop visual tests', () => {
             button = new Button();
             button.withQuery('knob-Button Type', 'primary');
             button.withQuery('knob-Button Size', 'small');
-            const pageUrl = buildUrl(button.componentType, button.componentName, button.path);
-
+            
             // Act
-            button.open(pageUrl);
-            button.waitForActionComponent();
+            button.load();
 
             // Assert
             browser.percyScreenshot('f-button - Primary - Small', 'desktop');
@@ -55,11 +48,9 @@ describe('f-button Desktop visual tests', () => {
             button = new Button();
             button.withQuery('knob-Button Type', 'primary');
             button.withQuery('knob-Button Size', 'xsmall');
-            const pageUrl = buildUrl(button.componentType, button.componentName, button.path);
-
+            
             // Act
-            button.open(pageUrl);
-            button.waitForActionComponent();
+            button.load();
 
             // Assert
             browser.percyScreenshot('f-button - Primary - XSmall', 'desktop');
@@ -72,11 +63,9 @@ describe('f-button Desktop visual tests', () => {
                 button.withQuery('knob-Button Type', 'primary');
                 button.withQuery('knob-Button Size', 'medium');
                 button.withQuery('knob-isLoading', 'true');
-                const pageUrl = buildUrl(button.componentType, button.componentName, button.path);
-
+                
                 // Act
-                button.open(pageUrl);
-                button.waitForActionComponent();
+                button.load();
 
                 // Assert
                 browser.percyScreenshot('f-button - Primary - Medium - Loading', 'desktop');
@@ -88,11 +77,9 @@ describe('f-button Desktop visual tests', () => {
                 button.withQuery('knob-Button Type', 'primary');
                 button.withQuery('knob-Button Size', 'large');
                 button.withQuery('knob-isLoading', 'true');
-                const pageUrl = buildUrl(button.componentType, button.componentName, button.path);
-
+                
                 // Act
-                button.open(pageUrl);
-                button.waitForActionComponent();
+                button.load();
 
                 // Assert
                 browser.percyScreenshot('f-button - Primary - Large - Loading', 'desktop');
@@ -106,11 +93,9 @@ describe('f-button Desktop visual tests', () => {
             button = new Button();
             button.withQuery('knob-Button Type', 'secondary');
             button.withQuery('knob-Button Size', 'medium');
-            const pageUrl = buildUrl(button.componentType, button.componentName, button.path);
-
+            
             // Act
-            button.open(pageUrl);
-            button.waitForActionComponent();
+            button.load();
 
             // Assert
             browser.percyScreenshot('f-button - Secondary - Medium', 'desktop');
@@ -121,11 +106,9 @@ describe('f-button Desktop visual tests', () => {
             button = new Button();
             button.withQuery('knob-Button Type', 'secondary');
             button.withQuery('knob-Button Size', 'large');
-            const pageUrl = buildUrl(button.componentType, button.componentName, button.path);
-
+            
             // Act
-            button.open(pageUrl);
-            button.waitForActionComponent();
+            button.load();
 
             // Assert
             browser.percyScreenshot('f-button - Secondary - Large', 'desktop');
@@ -136,11 +119,9 @@ describe('f-button Desktop visual tests', () => {
             button = new Button();
             button.withQuery('knob-Button Type', 'secondary');
             button.withQuery('knob-Button Size', 'small');
-            const pageUrl = buildUrl(button.componentType, button.componentName, button.path);
-
+            
             // Act
-            button.open(pageUrl);
-            button.waitForActionComponent();
+            button.load();
 
             // Assert
             browser.percyScreenshot('f-button - Secondary - Small', 'desktop');
@@ -151,11 +132,9 @@ describe('f-button Desktop visual tests', () => {
             button = new Button();
             button.withQuery('knob-Button Type', 'secondary');
             button.withQuery('knob-Button Size', 'xsmall');
-            const pageUrl = buildUrl(button.componentType, button.componentName, button.path);
-
+            
             // Act
-            button.open(pageUrl);
-            button.waitForActionComponent();
+            button.load();
 
             // Assert
             browser.percyScreenshot('f-button - Secondary - XSmall', 'desktop');
@@ -168,11 +147,9 @@ describe('f-button Desktop visual tests', () => {
                 button.withQuery('knob-Button Type', 'secondary');
                 button.withQuery('knob-Button Size', 'medium');
                 button.withQuery('knob-isLoading', 'true');
-                const pageUrl = buildUrl(button.componentType, button.componentName, button.path);
-
+                
                 // Act
-                button.open(pageUrl);
-                button.waitForActionComponent();
+                button.load();
 
                 // Assert
                 browser.percyScreenshot('f-button - Secondary - Medium - Loading', 'desktop');
@@ -184,11 +161,9 @@ describe('f-button Desktop visual tests', () => {
                 button.withQuery('knob-Button Type', 'secondary');
                 button.withQuery('knob-Button Size', 'large');
                 button.withQuery('knob-isLoading', 'true');
-                const pageUrl = buildUrl(button.componentType, button.componentName, button.path);
-
+                
                 // Act
-                button.open(pageUrl);
-                button.waitForActionComponent();
+                button.load();
 
                 // Assert
                 browser.percyScreenshot('f-button - Secondary - Large - Loading', 'desktop');
@@ -202,11 +177,9 @@ describe('f-button Desktop visual tests', () => {
             button = new Button();
             button.withQuery('knob-Button Type', 'outline');
             button.withQuery('knob-Button Size', 'medium');
-            const pageUrl = buildUrl(button.componentType, button.componentName, button.path);
-
+            
             // Act
-            button.open(pageUrl);
-            button.waitForActionComponent();
+            button.load();
 
             // Assert
             browser.percyScreenshot('f-button - Outline - Medium', 'desktop');
@@ -217,11 +190,9 @@ describe('f-button Desktop visual tests', () => {
             button = new Button();
             button.withQuery('knob-Button Type', 'outline');
             button.withQuery('knob-Button Size', 'large');
-            const pageUrl = buildUrl(button.componentType, button.componentName, button.path);
-
+            
             // Act
-            button.open(pageUrl);
-            button.waitForActionComponent();
+            button.load();
 
             // Assert
             browser.percyScreenshot('f-button - Outline - Large', 'desktop');
@@ -232,11 +203,9 @@ describe('f-button Desktop visual tests', () => {
             button = new Button();
             button.withQuery('knob-Button Type', 'outline');
             button.withQuery('knob-Button Size', 'small');
-            const pageUrl = buildUrl(button.componentType, button.componentName, button.path);
-
+            
             // Act
-            button.open(pageUrl);
-            button.waitForActionComponent();
+            button.load();
 
             // Assert
             browser.percyScreenshot('f-button - Outline - Small', 'desktop');
@@ -247,11 +216,9 @@ describe('f-button Desktop visual tests', () => {
             button = new Button();
             button.withQuery('knob-Button Type', 'outline');
             button.withQuery('knob-Button Size', 'xsmall');
-            const pageUrl = buildUrl(button.componentType, button.componentName, button.path);
-
+            
             // Act
-            button.open(pageUrl);
-            button.waitForActionComponent();
+            button.load();
 
             // Assert
             browser.percyScreenshot('f-button - Outline - XSmall', 'desktop');
@@ -264,11 +231,9 @@ describe('f-button Desktop visual tests', () => {
                 button.withQuery('knob-Button Type', 'outline');
                 button.withQuery('knob-Button Size', 'medium');
                 button.withQuery('knob-isLoading', 'true');
-                const pageUrl = buildUrl(button.componentType, button.componentName, button.path);
-
+                
                 // Act
-                button.open(pageUrl);
-                button.waitForActionComponent();
+                button.load();
 
                 // Assert
                 browser.percyScreenshot('f-button - Outline - Medium - Loading', 'desktop');
@@ -280,11 +245,9 @@ describe('f-button Desktop visual tests', () => {
                 button.withQuery('knob-Button Type', 'outline');
                 button.withQuery('knob-Button Size', 'large');
                 button.withQuery('knob-isLoading', 'true');
-                const pageUrl = buildUrl(button.componentType, button.componentName, button.path);
-
+                
                 // Act
-                button.open(pageUrl);
-                button.waitForActionComponent();
+                button.load();
 
                 // Assert
                 browser.percyScreenshot('f-button - Outline - Large - Loading', 'desktop');
@@ -298,11 +261,9 @@ describe('f-button Desktop visual tests', () => {
             button = new Button();
             button.withQuery('knob-Button Type', 'ghost');
             button.withQuery('knob-Button Size', 'medium');
-            const pageUrl = buildUrl(button.componentType, button.componentName, button.path);
-
+            
             // Act
-            button.open(pageUrl);
-            button.waitForActionComponent();
+            button.load();
 
             // Assert
             browser.percyScreenshot('f-button - Ghost - Medium', 'desktop');
@@ -313,11 +274,9 @@ describe('f-button Desktop visual tests', () => {
             button = new Button();
             button.withQuery('knob-Button Type', 'ghost');
             button.withQuery('knob-Button Size', 'large');
-            const pageUrl = buildUrl(button.componentType, button.componentName, button.path);
-
+            
             // Act
-            button.open(pageUrl);
-            button.waitForActionComponent();
+            button.load();
 
             // Assert
             browser.percyScreenshot('f-button - Ghost - Large', 'desktop');
@@ -328,11 +287,9 @@ describe('f-button Desktop visual tests', () => {
             button = new Button();
             button.withQuery('knob-Button Type', 'ghost');
             button.withQuery('knob-Button Size', 'small');
-            const pageUrl = buildUrl(button.componentType, button.componentName, button.path);
-
+            
             // Act
-            button.open(pageUrl);
-            button.waitForActionComponent();
+            button.load();
 
             // Assert
             browser.percyScreenshot('f-button - Ghost - Small', 'desktop');
@@ -343,11 +300,9 @@ describe('f-button Desktop visual tests', () => {
             button = new Button();
             button.withQuery('knob-Button Type', 'ghost');
             button.withQuery('knob-Button Size', 'xsmall');
-            const pageUrl = buildUrl(button.componentType, button.componentName, button.path);
-
+            
             // Act
-            button.open(pageUrl);
-            button.waitForActionComponent();
+            button.load();
 
             // Assert
             browser.percyScreenshot('f-button - Ghost - XSmall', 'desktop');
@@ -360,11 +315,9 @@ describe('f-button Desktop visual tests', () => {
                 button.withQuery('knob-Button Type', 'ghost');
                 button.withQuery('knob-Button Size', 'medium');
                 button.withQuery('knob-isLoading', 'true');
-                const pageUrl = buildUrl(button.componentType, button.componentName, button.path);
-
+                
                 // Act
-                button.open(pageUrl);
-                button.waitForActionComponent();
+                button.load();
 
                 // Assert
                 browser.percyScreenshot('f-button - Ghost - Medium - Loading', 'desktop');
@@ -376,11 +329,9 @@ describe('f-button Desktop visual tests', () => {
                 button.withQuery('knob-Button Type', 'ghost');
                 button.withQuery('knob-Button Size', 'large');
                 button.withQuery('knob-isLoading', 'true');
-                const pageUrl = buildUrl(button.componentType, button.componentName, button.path);
-
+                
                 // Act
-                button.open(pageUrl);
-                button.waitForActionComponent();
+                button.load();
 
                 // Assert
                 browser.percyScreenshot('f-button - Ghost - Large - Loading', 'desktop');
@@ -394,11 +345,9 @@ describe('f-button Desktop visual tests', () => {
             button = new Button();
             button.withQuery('knob-Button Type', 'link');
             button.withQuery('knob-Button Size', 'medium');
-            const pageUrl = buildUrl(button.componentType, button.componentName, button.path);
-
+            
             // Act
-            button.open(pageUrl);
-            button.waitForActionComponent();
+            button.load();
 
             // Assert
             browser.percyScreenshot('f-button - Link - Medium', 'desktop');
@@ -409,11 +358,9 @@ describe('f-button Desktop visual tests', () => {
             button = new Button();
             button.withQuery('knob-Button Type', 'link');
             button.withQuery('knob-Button Size', 'large');
-            const pageUrl = buildUrl(button.componentType, button.componentName, button.path);
-
+            
             // Act
-            button.open(pageUrl);
-            button.waitForActionComponent();
+            button.load();
 
             // Assert
             browser.percyScreenshot('f-button - Link - Large', 'desktop');
@@ -424,11 +371,9 @@ describe('f-button Desktop visual tests', () => {
             button = new Button();
             button.withQuery('knob-Button Type', 'link');
             button.withQuery('knob-Button Size', 'small');
-            const pageUrl = buildUrl(button.componentType, button.componentName, button.path);
-
+            
             // Act
-            button.open(pageUrl);
-            button.waitForActionComponent();
+            button.load();
 
             // Assert
             browser.percyScreenshot('f-button - Link - Small', 'desktop');
@@ -439,11 +384,9 @@ describe('f-button Desktop visual tests', () => {
             button = new Button();
             button.withQuery('knob-Button Type', 'link');
             button.withQuery('knob-Button Size', 'xsmall');
-            const pageUrl = buildUrl(button.componentType, button.componentName, button.path);
-
+            
             // Act
-            button.open(pageUrl);
-            button.waitForActionComponent();
+            button.load();
 
             // Assert
             browser.percyScreenshot('f-button - Link - XSmall', 'desktop');
@@ -456,11 +399,9 @@ describe('f-button Desktop visual tests', () => {
                 button.withQuery('knob-Button Type', 'link');
                 button.withQuery('knob-Button Size', 'medium');
                 button.withQuery('knob-isLoading', 'true');
-                const pageUrl = buildUrl(button.componentType, button.componentName, button.path);
-
+                
                 // Act
-                button.open(pageUrl);
-                button.waitForActionComponent();
+                button.load();
 
                 // Assert
                 browser.percyScreenshot('f-button - Link - Medium - Loading', 'desktop');
@@ -472,11 +413,9 @@ describe('f-button Desktop visual tests', () => {
                 button.withQuery('knob-Button Type', 'link');
                 button.withQuery('knob-Button Size', 'large');
                 button.withQuery('knob-isLoading', 'true');
-                const pageUrl = buildUrl(button.componentType, button.componentName, button.path);
-
+                
                 // Act
-                button.open(pageUrl);
-                button.waitForActionComponent();
+                button.load();
 
                 // Assert
                 browser.percyScreenshot('f-button - Link - Large - Loading', 'desktop');