+++ packages/components/molecules/f-media-element/test-utils/component-objects/f-mediaElement.component.js
@@ -1,4 +1,5 @@
 const Page = require('@justeat/f-wdio-utils/src/page.object');
+const { buildUrl } = require('@justeat/f-wdio-utils/src/storybook-extensions');
 
 module.exports = class MediaElement extends Page {
     constructor(){
@@ -6,6 +7,12 @@ module.exports = class MediaElement extends Page {
     }
 
     get component () { return $('[data-test-id="mediaElement-component"]'); }
+
+    load () {
+        const pageUrl = buildUrl(this.componentType, this.componentName, this.path);
+        this.open(pageUrl);
+        this.waitForComponent();
+    }
 
     open (url) {
         super.open(url);