+++ packages/components/molecules/f-user-message/test/accessibility/axe-accessibility.spec.js
@@ -1,4 +1,3 @@
-const { buildUrl } = require('@justeat/f-wdio-utils/src/storybook-extensions');
 const { getAccessibilityTestResults } = require('../../../../../../test/utils/axe-helper');
 
 const UserMessage = require('../../test-utils/component-objects/f-user-message.component');
@@ -7,9 +6,7 @@ const userMessage = new UserMessage();
 
 describe('Accessibility tests', () => {
     beforeEach(() => {
-        const pageUrl = buildUrl(userMessage.componentType, userMessage.componentName, userMessage.path);
-        userMessage.open(pageUrl);
-        userMessage.waitForComponent();
+        userMessage.load();
     });
 
     it('a11y - should test f-user-message component WCAG compliance', () => {