+++ packages/components/organisms/f-checkout/test/accessibility/axe-accessibility.spec.js
@@ -1,6 +1,5 @@
 import forEach from 'mocha-each';
 
-const { buildUrl } = require('@justeat/f-wdio-utils/src/storybook-extensions');
 const { getAccessibilityTestResults } = require('../../../../../../test/utils/axe-helper');
 
 const Checkout = require('../../test-utils/component-objects/f-checkout.component');
@@ -14,9 +13,7 @@ describe('Accessibility tests', () => {
         checkout.withQuery('&knob-Service Type', 'delivery')
             .withQuery('&knob-Is User Logged In', true);
 
-        const pageUrl = buildUrl(checkout.componentType, checkout.componentName, checkout.path);
-        checkout.open(pageUrl);
-        checkout.waitForComponent();
+        checkout.load();
         const axeResults = getAccessibilityTestResults('f-checkout-delivery');
 
         // Assert
@@ -29,9 +26,7 @@ describe('Accessibility tests', () => {
         checkout.withQuery('&knob-Service Type', 'collection')
             .withQuery('&knob-Is User Logged In', true);
 
-        const pageUrl = buildUrl(checkout.componentType, checkout.componentName, checkout.path);
-        checkout.open(pageUrl);
-        checkout.waitForComponent();
+        checkout.load();
         const axeResults = getAccessibilityTestResults('f-checkout-collection');
 
         // Assert
@@ -44,9 +39,7 @@ describe('Accessibility tests', () => {
         checkout.withQuery('&knob-Service Type', 'delivery')
             .withQuery('&knob-Is User Logged In', false);
 
-        const pageUrl = buildUrl(checkout.componentType, checkout.componentName, checkout.path);
-        checkout.open(pageUrl);
-        checkout.waitForComponent();
+        checkout.load();
         const axeResults = getAccessibilityTestResults('f-checkout-guest');
 
         // Assert
@@ -59,9 +52,7 @@ describe('Accessibility tests', () => {
         checkout.withQuery('&knob-Service Type', 'Invalid URL')
             .withQuery('&knob-Is User Logged In', false);
 
-        const pageUrl = buildUrl(checkout.componentType, checkout.componentName, checkout.path);
-        checkout.open(pageUrl);
-        checkout.waitForErrorPageComponent();
+        checkout.load('error');
         const axeResults = getAccessibilityTestResults('f-checkout-error-page');
 
         // Assert
@@ -86,15 +77,14 @@ describe('Accessibility tests', () => {
             .withQuery('&knob-Is User Logged In', true)
             .withQuery('&knob-Patch Checkout Errors', patchError);
 
-        const pageUrl = buildUrl(checkout.componentType, checkout.componentName, checkout.path);
-
         // Act
-        checkout.open(pageUrl);
-        checkout.waitForComponent();
+        checkout.load();
         checkout.goToPayment();
 
+        const expectedTabOrder = [checkout.errorMessageRetry, checkout.closeMessageModal];
+        const result = checkout.testTabOrder(expectedTabOrder);
+
         // Assert
-        const result = checkout.getCheckoutErrorTabOrder();
         expect(result.actual).toEqual(result.expected);
     });
 
@@ -109,15 +99,14 @@ describe('Accessibility tests', () => {
             .withQuery('&knob-Is User Logged In', true)
             .withQuery('&knob-Place Order Errors', 'duplicate');
 
-        const pageUrl = buildUrl(checkout.componentType, checkout.componentName, checkout.path);
-
         // Act
-        checkout.open(pageUrl);
-        checkout.waitForComponent();
+        checkout.load();
         checkout.goToPayment();
 
+        const expectedTabOrder = [checkout.errorMessageRetry, checkout.errorMessageDupOrderGoToHistory, checkout.closeMessageModal];
+        const result = checkout.testTabOrder(expectedTabOrder);
+
         // Assert
-        const result = checkout.getPlaceOrderErrorTabOrder();
         expect(result.actual).toEqual(result.expected);
     });
 
@@ -133,9 +122,7 @@ describe('Accessibility tests', () => {
             .withQuery('&knob-Is User Logged In', false)
             .withQuery('&knob-Patch Checkout Errors', patchError);
 
-        const pageUrl = buildUrl(checkout.componentType, checkout.componentName, checkout.path);
-        checkout.open(pageUrl);
-        checkout.waitForComponent();
+        checkout.load();
 
         // Act
         checkout.setFieldValue('firstName', 'Jerry');
@@ -147,8 +134,10 @@ describe('Accessibility tests', () => {
         checkout.populateGuestCollectionCheckoutForm(addressInfo);
         checkout.goToPayment();
 
+        const expectedTabOrder = [checkout.errorMessageRetry, checkout.closeMessageModal];
+        const result = checkout.testTabOrder(expectedTabOrder);
+
         // Assert
-        const result = checkout.getCheckoutErrorTabOrder();
         expect(result.actual).toEqual(result.expected);
     });
 
@@ -164,9 +153,7 @@ describe('Accessibility tests', () => {
             .withQuery('&knob-Is User Logged In', false)
             .withQuery('&knob-Patch Checkout Errors', patchError);
 
-        const pageUrl = buildUrl(checkout.componentType, checkout.componentName, checkout.path);
-        checkout.open(pageUrl);
-        checkout.waitForComponent();
+        checkout.load();
 
         // Act
         checkout.setFieldValue('firstName', 'Jerry');
@@ -181,8 +168,10 @@ describe('Accessibility tests', () => {
         checkout.populateGuestCheckoutForm(addressInfo);
         checkout.goToPayment();
 
+        const expectedTabOrder = [checkout.errorMessageRetry, checkout.closeMessageModal];
+        const result = checkout.testTabOrder(expectedTabOrder);
+
         // Assert
-        const result = checkout.getCheckoutErrorTabOrder();
         expect(result.actual).toEqual(result.expected);
     });
 
@@ -198,9 +187,7 @@ describe('Accessibility tests', () => {
             .withQuery('&knob-Is User Logged In', false)
             .withQuery('&knob-Patch Checkout Errors', patchError);
 
-        const pageUrl = buildUrl(checkout.componentType, checkout.componentName, checkout.path);
-        checkout.open(pageUrl);
-        checkout.waitForComponent();
+        checkout.load();
 
         // Act
         checkout.setFieldValue('firstName', 'Jerry');
@@ -213,8 +200,10 @@ describe('Accessibility tests', () => {
         checkout.populateGuestDineInCheckoutForm(addressInfo);
         checkout.goToPayment();
 
+        const expectedTabOrder = [checkout.errorMessageRetry, checkout.closeMessageModal];
+        const result = checkout.testTabOrder(expectedTabOrder);
+
         // Assert
-        const result = checkout.getCheckoutErrorTabOrder();
         expect(result.actual).toEqual(result.expected);
     });
 
@@ -225,9 +214,7 @@ describe('Accessibility tests', () => {
             .withQuery('&knob-Is User Logged In', false)
             .withQuery('&knob-Place Order Errors', 'duplicate');
 
-        const pageUrl = buildUrl(checkout.componentType, checkout.componentName, checkout.path);
-        checkout.open(pageUrl);
-        checkout.waitForComponent();
+        checkout.load();
 
         // Act
         checkout.setFieldValue('firstName', 'Jerry');
@@ -239,8 +226,10 @@ describe('Accessibility tests', () => {
         checkout.populateGuestCollectionCheckoutForm(addressInfo);
         checkout.goToPayment();
 
+        const expectedTabOrder = [checkout.errorMessageRetry, checkout.errorMessageDupOrderGoToHistory, checkout.closeMessageModal];
+        const result = checkout.testTabOrder(expectedTabOrder);
+
         // Assert
-        const result = checkout.getPlaceOrderErrorTabOrder();
         expect(result.actual).toEqual(result.expected);
     });
 
@@ -251,9 +240,7 @@ describe('Accessibility tests', () => {
             .withQuery('&knob-Is User Logged In', false)
             .withQuery('&knob-Place Order Errors', 'duplicate');
 
-        const pageUrl = buildUrl(checkout.componentType, checkout.componentName, checkout.path);
-        checkout.open(pageUrl);
-        checkout.waitForComponent();
+        checkout.load();
 
         // Act
         checkout.setFieldValue('firstName', 'Jerry');
@@ -268,8 +255,10 @@ describe('Accessibility tests', () => {
         checkout.populateGuestCheckoutForm(addressInfo);
         checkout.goToPayment();
 
+        const expectedTabOrder = [checkout.errorMessageRetry, checkout.errorMessageDupOrderGoToHistory, checkout.closeMessageModal];
+        const result = checkout.testTabOrder(expectedTabOrder);
+
         // Assert
-        const result = checkout.getPlaceOrderErrorTabOrder();
         expect(result.actual).toEqual(result.expected);
     });
 
@@ -280,9 +269,7 @@ describe('Accessibility tests', () => {
             .withQuery('&knob-Is User Logged In', false)
             .withQuery('&knob-Place Order Errors', 'duplicate');
 
-        const pageUrl = buildUrl(checkout.componentType, checkout.componentName, checkout.path);
-        checkout.open(pageUrl);
-        checkout.waitForComponent();
+        checkout.load();
 
         // Act
         checkout.setFieldValue('firstName', 'Jerry');
@@ -295,8 +282,10 @@ describe('Accessibility tests', () => {
         checkout.populateGuestDineInCheckoutForm(addressInfo);
         checkout.goToPayment();
 
+        const expectedTabOrder = [checkout.errorMessageRetry, checkout.errorMessageDupOrderGoToHistory, checkout.closeMessageModal];
+        const result = checkout.testTabOrder(expectedTabOrder);
+
         // Assert
-        const result = checkout.getPlaceOrderErrorTabOrder();
         expect(result.actual).toEqual(result.expected);
     });
 });