+++ packages/components/organisms/f-header/test/component/f-header.component.shared.spec.js
@@ -1,5 +1,3 @@
-const { buildUrl } = require('@justeat/f-wdio-utils/src/storybook-extensions');
-
 const Header = require('../../test-utils/component-objects/f-header.component');
 
 let header;
@@ -10,11 +8,9 @@ describe('Shared - f-header component tests - @percy', () => {
         header.withQuery('&knob-Locale', 'en-GB');
         header.withQuery('&knob-Show offers link', 'true');
         header.withQuery('&knob-Show delivery enquiry', 'true');
-        const pageUrl = buildUrl(header.componentType, header.componentName, header.path);
 
         // Act
-        header.open(pageUrl);
-        header.waitForComponent();
+        header.load();
     });
 
     it('should display component', () => {