+++ CHANGELOG.md
@@ -4,6 +4,26 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 
+v4.0.0
+------------------------------
+*September 15, 2021*
+
+### Updated
+- fozzie version to make normalize styles optional. This will remove normalize styles from the components by default.
+- JETSansDigital font instead of JustEatBasis for the components and storybook.
+
+
+v4.0.0-beta.0
+------------------------------
+*August 25, 2021*
+
+### Updated
+- fozzie version to include new font.
+
+### Changed
+- Storybook font to become new JETSansDigital instead of JustEatBasis.
+
+
 v3.54.0
 ------------------------------
 *September 2, 2021*