+++ packages/components/molecules/f-mega-modal/CHANGELOG.md
@@ -3,12 +3,40 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+
+v2.0.0
+------------------------------
+*September 16, 2021*
+
+### Changed
+- Updated version of `f-button`.
+
+### Removed
+- Normalise styles from the build. Note that now if consuming application doesn't have normalised (reset) styles, there may be some slight style differences after this update. If the consuming application uses fozzie, there shouldn't be any style differences as normalised styles should be added as part of the fozzie import.
+
+
+v1.0.0
+------------------------------
+*September 15, 2021*
+
+- Return beta to master. Component has JETSansDigital font.
+
+
+v1.0.0-beta.0
+------------------------------
+*September 1, 2021*
+
+### Changed
+- Updated version of `f-button` (JETSans Update)
+
+
 v0.12.1
 ------------------------------
 *September 1, 2021*
 
 ### Changed
 - Fixed bug where scrolling is not re-enabled if the modal is destroyed while open for whatever reason.
+
 
 v0.12.0
 ------------------------------