+++ packages/components/organisms/f-offers/CHANGELOG.md
@@ -4,6 +4,46 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 
+v1.1.0
+------------------------------
+*September 16, 2021*
+
+### Changed
+- Updated version of `f-button`, `f-content-cards`, `f-media-element`, `f-searchbox` and `f-mega-modal` to remove normalize styles from the build.
+
+
+v1.0.0
+------------------------------
+*September 15, 2021*
+
+- Return beta to master. Component has JETSansDigital font.
+
+
+v1.0.0-beta.2
+------------------------------
+*September 7, 2021*
+
+### Changed
+- Updated version of `f-searchbox` (JETSans Update).
+
+
+v1.0.0-beta.1
+------------------------------
+*September 6, 2021*
+
+### Changed
+- Updated version of `f-mega-modal` (JETSans Update).
+
+
+v1.0.0-beta.0
+------------------------------
+*August 26, 2021*
+
+### Changed
+- New font JETSansDigital
+- `f-button` to v2.0.0-beta.0 which use the new font
+
+
 v0.8.0
 ------------------------------
 *August 11, 2021*